requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12
djangorestframework-simplejwt>=4.7
drf-yasg>=1.20
setuptools>=75.1.0